Move r.v.interfaces package to r.views package
