+++ app/src/main/java/com/shapeapp/shape/ui/publicpackage/PublicFragment.kt
@@ -17,7 +17,7 @@ import com.shapeapp.shape.data.entities.Card
 import com.shapeapp.shape.ui.dialogs.MyPublicSharesBSD
 import com.shapeapp.shape.mockupsmakers.UsersMockups
 import com.shapeapp.shape.ui.recyclerviews.recyclerviewadapters.SmallCardRecyclerViewAdapter
-import com.shapeapp.shape.recyclerviewinterfaces.RecyclerViewCardClickListener
+import com.shapeapp.shape.ui.recyclerviews.recyclerviewinterfaces.RecyclerViewCardClickListener
 import com.shapeapp.shape.repositories.CardsRepository
 import kotlinx.android.synthetic.main.fragment_public.*
